code cleanup around mir_snprintf
git-svn-id: http://svn.miranda-ng.org/main/trunk@11285 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Scriver/src/msgs.cpp
@@ -299,7 +299,7 @@ static INT_PTR GetWindowAPI(WPARAM wParam, LPARAM lParam)
 static INT_PTR GetWindowClass(WPARAM wParam, LPARAM lParam)
 {
 	char *szBuf = (char*)wParam;
-	int size = (int)lParam;
+	size_t size = (size_t)lParam;
 	mir_snprintf(szBuf, size, "Scriver");
 	return 0;
 }
